Add files via upload
Since Packet Tracer simulates networks and network traffic, the physical aspects of these networks also needs to be simulated. This includes actually finding and deploying physical devices, customizing those devices, and cabling those devices. After the physical deployment and cabling is done, then it is time for configuration of the interfaces used to connect the devices.